+++ src/Rule/Valid.php
@@ -48,14 +48,32 @@ final class Valid
     public function get_validation(): string
     {
         $is_invert       = false;
+        $is_block_if     = false;
         $validation_rule = [];
 
         foreach ($this->validation_rule as $rule) {
-            if ($rule == 'invert') {
+            // detect if condition
+            if ($rule === 'if_false') {
+                $is_block_if = true;
+                continue;
+            }
+            if ($rule === 'if_true' || $rule === 'end_if') {
+                // break if statment
+                $is_block_if = false;
+                continue;
+            }
+            // block rule if statment is false
+            if ($is_block_if === true) {
+                continue;
+            }
+
+            // set next rule as invert rule
+            if ($rule === 'invert') {
                 $is_invert = !$is_invert;
                 continue;
             }
 
+            // add string rule and reset invert rule
             $validation_rule[] = $is_invert ? 'invert_' . $rule : $rule;
             $is_invert         = false;
         }
@@ -103,24 +121,60 @@ final class Valid
      *
      * Reset only boolean false.
      *
-     * @param Closure $condition Closure return boolean (default false)
+     * @param Closure $condition Closure return boolean
      */
     public function where(Closure $condition): string
     {
-        // default is false
+        // get return closure
         $result = call_user_func_array($condition, []);
-
+        // throw exception if closure not return boolean
         if (!is_bool($result)) {
             throw new Exception('Condition closure not return boolean', 1);
         }
 
-        // false condtion and not bool return will reset Rule
+        // false condition
         if ($result === false) {
             $this->validation_rule = [];
         }
 
         // prevent create new rule and give a string rule
         return $this->get_validation();
+    }
+
+    /**
+     * Rule will be applay if condition is true,
+     * otherwise rule be skip if return false.
+     *
+     * Reset only boolean false.
+     *
+     * @param Closure $condition Closure return boolean
+     */
+    public function if(Closure $condition): self
+    {
+        // get return closure
+        $result = call_user_func_array($condition, []);
+        // throw exception if closure not return boolean
+        if (!is_bool($result)) {
+            throw new Exception('Condition closure not return boolean', 1);
+        }
+
+        // add condition to rule
+        $this->validation_rule[] = $result
+            ? 'if_true'
+            : 'if_false'
+        ;
+
+        return $this;
+    }
+
+    /**
+     * Set end rule of 'if' statment.
+     */
+    public function end_if(): self
+    {
+        $this->validation_rule[] = 'end_if';
+
+        return $this;
     }
 
     // -------------------------------------------------------